backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-nmap>=0.7.1
